+++ generators/dotnetcore-server-interfaces/OpenApiCodegen.Server.Mvc.Base/CSharpControllerTransformer.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using DarkPatterns.OpenApiCodegen;
+using DarkPatterns.OpenApiCodegen.Handlebars;
 
 namespace DarkPatterns.OpenApi.CSharp;
