+++ generators/dotnetcore-server-interfaces/OpenApiCodegen.Server.Mvc.Base/Templates/AddServices.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using DarkPatterns.OpenApiCodegen.Handlebars.Templates;
 
-namespace DarkPatterns.OpenApi.CSharp.Templates
-{
-	public record AddServicesModel(Templates.PartialHeader Header, string MethodName, string PackageName, ControllerReference[] Controllers);
+namespace DarkPatterns.OpenApi.CSharp.Templates;
 
-	public record ControllerReference(string GenericTypeName, string ClassName);
-}
+public record AddServicesModel(PartialHeader Header, string MethodName, string PackageName, ControllerReference[] Controllers);
+
+public record ControllerReference(string GenericTypeName, string ClassName);
