+++ generators/dotnetstandard-client/OpenApiCodegen.Client.Base/ClientTransformerFactory.cs
@@ -1,6 +1,8 @@
 ﻿using DarkPatterns.Json.Documents;
 using DarkPatterns.OpenApi.Transformations;
 using DarkPatterns.OpenApi.Abstractions;
+using DarkPatterns.OpenApiCodegen.Handlebars;
+
 namespace DarkPatterns.OpenApi.CSharp;
 
 public class ClientTransformerFactory(TransformSettings settings)
