+++ generators/typescript/OpenApiCodegen.Client.TypeScript/OpenApiCodegen.Client.TypeScript.csproj
@@ -24,5 +24,6 @@
 	<ItemGroup>
 		<ProjectReference Include="$(SolutionRoot)lib\Extensions.Configuration.Yaml\Extensions.Configuration.Yaml.csproj" SpecificVersion="true" />
 		<ProjectReference Include="$(SolutionRoot)lib\OpenApi.TypeScript\OpenApi.TypeScript.csproj" />
+		<ProjectReference Include="..\..\..\lib\OpenApiCodegen.Handlebars\OpenApiCodegen.Handlebars.csproj" />
 	</ItemGroup>
 </Project>
