+++ lib/OpenApi.CSharp/HandlebarsFactoryDefaults.cs
@@ -1,0 +1,9 @@
+﻿using DarkPatterns.OpenApiCodegen.Handlebars;
+using HandlebarsDotNet;
+
+namespace DarkPatterns.OpenApi.CSharp;
+
+public static class HandlebarsFactoryDefaults
+{
+	public static HandlebarsFactory Default { get; } = new(CSharpHandlebarsCommon.CreateHandlebars);
+}
