--- lib/OpenApi.CSharp/Templates/partial_header.handlebars
@@ -1,14 +1,0 @@
-/*
-{{#if AppName}}
- * {{{AppName}}} {{#if Version}}v{{{Version}}}{{/if}}
- *
-{{/if}}
-{{#if AppDescription}}
- * {{{AppDescription}}}
- *
-{{/if}}
- *
- * {{#if InfoEmail}}API Contact: {{{InfoEmail}}}{{/if}}
- *
- * This source generated by: https://principle.tools{{#if CodeGeneratorVersionInfo}} ({{{CodeGeneratorVersionInfo}}}){{/if}}
- */
