+++ lib/OpenApi.CSharp/TransformSettings.cs
@@ -2,10 +2,11 @@
 using DarkPatterns.OpenApi.Abstractions;
 using System.Linq;
 using DarkPatterns.OpenApi.Transformations;
+using DarkPatterns.OpenApiCodegen.Handlebars.Templates;
 
 namespace DarkPatterns.OpenApi.CSharp;
 
-public record TransformSettings(SchemaRegistry SchemaRegistry, Templates.PartialHeader Header)
+public record TransformSettings(SchemaRegistry SchemaRegistry, PartialHeader Header)
 {
 	// TODO: This doesn't seem like the right spot, but it is the correct namespace for the current level of abstraction... which may need to change.
 	public static CompositeOpenApiSourceProvider BuildComposite(
@@ -14,7 +15,7 @@ public record TransformSettings(SchemaRegistry SchemaRegistry, Templates.Partial
 		string versionInfo,
 		System.Func<TransformSettings, ISourceProvider>[] factories)
 	{
-		var header = new Templates.PartialHeader(
+		var header = new PartialHeader(
 			AppName: document.Info.Title,
 			AppDescription: document.Info.Description,
 			Version: document.Info.Version,
