+++ lib/OpenApi.TypeScript/OpenApi.TypeScript.csproj
@@ -4,20 +4,13 @@
 		<TargetFramework>netstandard2.0</TargetFramework>
 		<VersionPrefix>$(SharedAnalyzerLibrariesVersion)</VersionPrefix>
 	</PropertyGroup>
-	<ItemGroup>
-		<PackageReference Include="Handlebars.Net" Version="2.1.4" />
-	</ItemGroup>
 	<Import Project="$(RepositoryEngineeringDir)netstandard/polyfills.targets" />
 	<ItemGroup>
-		<EmbeddedResource Include="$(MSBuildThisFileDirectory)Templates\*.handlebars">
-			<LogicalName>$(RootNamespacePrefix).OpenApi.TypeScript.Templates.%(Filename)%(Extension)</LogicalName>
-		</EmbeddedResource>
-		<EmbeddedResource Include="$(MSBuildThisFileDirectory)typescript.config.yaml">
-			<CopyToOutputDirectory>Never</CopyToOutputDirectory>
-			<LogicalName>$(RootNamespacePrefix).OpenApi.TypeScript.%(Filename)%(Extension)</LogicalName>
-		</EmbeddedResource>
+		<EmbeddedResource Include="$(MSBuildThisFileDirectory)Templates\*.handlebars" />
+		<EmbeddedResource Include="$(MSBuildThisFileDirectory)typescript.config.yaml" />
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\OpenApi.Transformations\OpenApi.Transformations.csproj" SpecificVersion="true" />
+		<ProjectReference Include="..\OpenApiCodegen.Handlebars\OpenApiCodegen.Handlebars.csproj" SpecificVersion="true" />
 	</ItemGroup>
 </Project>
