--- lib/OpenApi.TypeScript/Templates/PartialHeader.cs
@@ -1,10 +1,0 @@
-﻿namespace DarkPatterns.OpenApi.TypeScript.Templates
-{
-	public record PartialHeader(
-		string? AppName,
-		string? AppDescription,
-		string? Version,
-		string? InfoEmail,
-		string CodeGeneratorVersionInfo
-	);
-}
