+++ lib/OpenApi.TypeScript/TransformSettings.cs
@@ -1,5 +1,6 @@
 ﻿using DarkPatterns.OpenApi.Transformations;
+using DarkPatterns.OpenApiCodegen.Handlebars.Templates;
 
 namespace DarkPatterns.OpenApi.TypeScript;
 
-public record TransformSettings(ISchemaRegistry SchemaRegistry, Templates.PartialHeader Header);
+public record TransformSettings(ISchemaRegistry SchemaRegistry, PartialHeader Header);
