+++ lib/OpenApiCodegen.Handlebars/OpenApiCodegen.Handlebars.csproj
@@ -4,14 +4,14 @@
 		<TargetFramework>netstandard2.0</TargetFramework>
 		<VersionPrefix>$(SharedAnalyzerLibrariesVersion)</VersionPrefix>
 	</PropertyGroup>
-	<Import Project="$(RepositoryEngineeringDir)netstandard/polyfills.targets" />
 	<ItemGroup>
-		<EmbeddedResource Include="Templates\*.handlebars" />
-		<EmbeddedResource Include="csharp.config.yaml" />
+		<PackageReference Include="Handlebars.Net" Version="2.1.4" />
 	</ItemGroup>
-
 	<ItemGroup>
 		<ProjectReference Include="..\OpenApi.Transformations\OpenApi.Transformations.csproj" SpecificVersion="true" />
-		<ProjectReference Include="..\OpenApiCodegen.handlebars\OpenApiCodegen.handlebars.csproj" SpecificVersion="true" />
 	</ItemGroup>
+	<ItemGroup>
+		<EmbeddedResource Include="Templates\*.handlebars" />
+	</ItemGroup>
+	<Import Project="$(RepositoryEngineeringDir)netstandard/polyfills.targets" />
 </Project>
