+++ lib/OpenApiCodegen.Handlebars/Templates/partial_header.handlebars
@@ -11,5 +11,5 @@
  * API Contact: {{{InfoEmail}}}
  *
 {{/if}}
- * This source generated by: https://principle.tools{{#if CodeGeneratorVersionInfo}} ({{{CodeGeneratorVersionInfo}}}){{/if}}
+ * This source generated by: https://tools.darkpatterns.com{{#if CodeGeneratorVersionInfo}} ({{{CodeGeneratorVersionInfo}}}){{/if}}
  */
